+++ spec/lib/pr/common/charge_service_spec.rb
@@ -38,7 +38,8 @@ describe PR::Common::ChargeService do
           user_id: user.id,
           traits: {
             monthlyUsd: charge.price,
-            appPlan: described_class.determine_app_plan_from_charge(charge)
+            appPlan: described_class.determine_app_plan_from_charge(charge),
+            trial: price.positive?
           }
         )